backend/pyproject.toml
[tool.poetry]
name = "musicmatch-backend"
version = "0.1.0"
description = "Backend for MusicMatch application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
pymongo = "^4.6.1"
python-dotenv = "^1.0.1"
django-ninja = "^1.0.1"
pydantic = "^2.6.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"